restructure project to resolve dependency cycle
