--- server/controllers/index.js
@@ -1,3 +1,0 @@
-module.exports.product = require('./product.controller');
-module.exports.style = require('./style.controller');
-module.exports.related = require('./related.controller');
