+++ server/index.js
@@ -2,13 +2,13 @@ require('dotenv').config();
 const express = require('express');
 // const path = require('path');
 const cors = require('cors');
+const router = require('./routes');
 
 const app = express();
 
 app.use(cors());
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-const router = require('./routes');
 
 app.use('/products', router);
